+++ f2837xd/chrono/chrono.cpp
@@ -45,7 +45,7 @@ void system_clock::init() {
 
 
 void system_clock::run_tasks() {
-    for (size_t i = 0; i < _tasks.size(); ++i) {
+    for (int i = 0; i < _tasks.size(); ++i) {
         if (now() >= (_tasks[i].timepoint + _tasks[i].period)) {
             if (_tasks[i].func(i) == TaskStatus::success) {
                 _tasks[i].timepoint = now();
